requirements.txt
Flask>=2.0.0,<3.0.0
requests>=2.25.0,<3.0.0
google-cloud-storage>=1.40.0,<3.0.0
gunicorn>=20.0.0,<22.0.0